added comments to clarify several functions
SXW_GetPR is no longer utilized in the current code scheme. So I added
a comment that reflects that.

Also, in ST_resgroups.c there are different chunks of code that
calculate res_required and res_avail when running SOILWAT and when not
running SOILWAT. I added some comments to make that clear.