Added "succulents" as a group
We added succulents back as a potential resource group. As such, the
other files needed to be updated. There are now 11 potential resource
groups, although only 10 can be specified and turned on at one time.

Also reset the rgroup.in file, so disturbances (fire and grazing) are
off as the default.