Updated sagebrush root distribution
Each row must sum to 1. Somewhere alone the line, this line got
altered, so the values do not match what Trace used in her simulation
runs and the values exceed 1. I updated the parameters to sum to 1 and
to reflect the values we used in Trace’s manuscript and simulation runs.