Replaced historical weather data and markov files
The old mkv_covar.in and mkv_prob.in files were not compatible with the
new changes to the the markov weather generator that Trace implemented
in STEPWAT on December 14th. I have added new mkv_covar.in and
mkv_prob.in files that are compatible with these changes to the C-code
and the site-specific historical data they were generated from. In the
future, the STEPWAT_R_WRAPPER can we use to generated new mkv_covar.in
and mkv_prob.in files for site-specific historical weather data. There
was no capacity to do this either in SOILWAT or STEPWAT.