Remove unnecessary constant
the no_resources function kills plants when soil water resources become
limiting. This function is based on equation 7 in Coffin and Lauenroth
1990. Somehow along the line a constant (0.5) got inserted into the
model code. Unfortunately there is no comment that justified this
change and modification of the original function in Coffin and
Lauenroth. K. Palmquist tested removing the 0.5 and retaining the 0.5
under warm,dry weather conditions, cold, wet conditions, and relatively
warm, wet conditions and found removing the constant did not impact the
biomass or mortality output very much at all. Since we have no
theoretical reason to keep the 0.5 and since it does not impact the
output, I have removed it from the no_resources function.