updated add_annual_seedprod function
Added a check if the current year is equal to year 1 of the simulation.
If so, then addition to the seed bank will be a function of a random
number draw. If it is not year 1, then it will be a function of last
year’s biomass. Also removed outdated comments and unused parameters.