New algorithm - dividing extra resources Issue#86
This commit removes old code that is used when SOILWAT2 is not running
and overhauls the resource partitioning functions so that “extra”
resources (base_rem) are calculated at the functional group level and
then across all functional groups (total extra resources). Then, the
code that divides those extra resources to functional groups and then
individuals was modified.