Resolving issue #92
This commit adds an additional loop in _get_annual_maxestab adding the seed decay process to affect the seed prod arrays rather than the variable viable_seeds.

I added a new local variable called viable_seeds_reduced, which is viable seeds - seeds added this year. Calculation of this was necessary to properly remove seeds that germinated from the seedbank. It was important to not remove seeds that were just added this year.

Finally, comments were clarified, print statements added for tracking, and the code was cleaned up.