Fixed small floating point errors specified in issue #111
To fix the problem of individual relsizes not summing to the species relsize, I "synced" them after extra growth is removed at the end of the year. In ST_mortality.c, instead of subtracting the extra growth from relsize for each species I reset species relsize to 0 then added all of the individual relsizes back. This passively removes extra growth and also ensures that any errors related to floating point are carried over to the species and R group levels.