Modifies use of get_running_sqr() (issue #103)
The running sum of squares function needs to be added to the current sum of squares every year instead of replacing it. This commit corrects the calculation of the running standard deviation. 

I also changed the threshold for adding transpired water to 1 standard deviation until we can determine what the true value should be.