Transpiration average calculated using moving window (issue #169)
transpiration average data now stored in its own struct. The struct contains information on where the oldest value is stored, what the average and sum of squares are, and two arrays with the old sum of squares and average transpirations.

This struct is now used in sxw_resource.c to determine if adding transpiration is necessary.