Fixed transp_window.average bug and documented code (issue #169)
In sxw.c:
Added explanation of "window" and fixed typo.

in sxw.h
removed last_ratio field. It is only needed when performing a running average and does not need to be stored in a struct.
Renamed window to ratios to make its purpose more clear. 
Added documentation on how to use the struct and the relationships between variables.
Added a transp[] array to store raw transpiration data. This is needed to calculate the average transpiration.

In sxw_resource.c:
average to transpiration now calculated along side transpiration ratio.
Commented the code to make it more clear.