fixed window initialization issue #169, commened out Beta distribution issue #174
Every new iteration STEPWAT runs a clean up cycle that resets variables. This clean up cycle results in _transp_contribution_by_group() being called twice during the last year of a given iteration. This was affecting the moving window. The function now checks to make sure that the same year is not repeated. If it is repeated, it resets the transp_window fields then returns.

The added transpiration code was resulting in negative transpiration being added. The problem is that if RandBeta() is very small then (1 - transp_ratio/RandBeta) becomes negative. For now, everything invloved in adding transpiration is commented out until a decision is reached.