Added checks and documentation to STdebug (issues #201 & #202)
The switch case statement in init_args() keys off of the first letter in the flag. From there, I added code to see if the entire flag reads "-STdebug". If so, the program executes and generates the database. If not, the code prints valid options and returns. Additionally, any flag that starts with "-S" will now give the user a hint: "(did you mean -STdebug?)"

In usage() I added the -STdebug option.