Resolved issue #207
We removed NGrpEstab (i.e. grpMaxEstab) from the global structure in master branch because it was redundant and not used. This variable was still present in both ST_params.c and at the top of the disturbance.in within the Fire_testing branch. This resulted in an error. This commit resolves the problem by simply removing all reference to those variables.