Made submodule updating instructions more generals (issue #211)
The instructions for changing development branches are now more general. They explain how to checkout the new branch, and how to update all submodules. If the user is switching between branches that have the same submodules the second line, `git submodule update --init --recursive`, will do nothing but it doesn't hurt to try.