(issue #209) Fixed relsize < 0
The fix I added is in Species_Update_Newsize. If species relsize < 0 we can assume that the sum of individual relsizes and species relsize are out of sync. Now, when the function is called it first checks to see if species relsize <0. if true, it then sums up individual relsizes and resets species relsize. If species relize is still <0 we know that there are individuals with a negative relsize, and we generate an appropriate warning. This fix removed the warnings from my tests, but a few more tests to confirm that it works would be a good idea.