Removed all references to UseSoilwat (Issue #77)
We now expect to always use SOILWAT2 to calculate resources. With this in mind, I removed the variable UseSoilwat. Wherever it was used, I treated the code as if UseSoilwat was TRUE, and modified it accordingly. 

Because we will always be using SOILWAT from now on, I also removed the -s option on startup. This is equivalent to always including the -s flag.