Prototyped CellType struct, created ST_stats.h
In ST_grid.c: I added the CellType struct with a few variables that I know need to be plot-specific. This struct is never instantiated in the code, it is only there to discuss and augment. 

In ST_stats.h: created ST_stats.h and moved all struct definitions from ST_stats.c to the new file. This way, we can instantiate accumulators in other files like ST_grid.c.

In ST_stats.c: Moved all struct definitions out of this file. This file now includes ST_stats.h and uses the struct definitions from it.