Compress sxw.c variables into struct.
I added a new struct to sxw.h that contains all of the local variables from sxw.c This way we can instantiate these variables in each grid cell when using gridded mode. 

I named this struct TempType for now until we agree on a new name. The implementation of this struct in sxw.c is called SXWTemp. I am commiting these changes so that I can move on to other issues, and will change the names once we come up with something descriptive.

I also added two local variables to the transp_window struct: added_transp and lastYear. These were free floating variables that need to be cell specific.