Addressing comment on pull request #270
The log warnings in Plot_Initialize needed to be updated to reflect the fact that relsize is now calculated on the fly every time you call the getRelsize function. The only way this warning could get printed now is if all individuals were not killed after an iteration.

I also removed the programmer alert and replaced it with a reference to the related issue on GitHub.