Moved printing general info to RunGrid
(issue #287)
SXW_Init was printing the non-gridded mode information to stdout. The problem with this was the fact that we call SXW_Init n times where n is the number of cells. This resulted in the same message being generated multiple times.
I wrapped the print statement in SXW_Init in an if(!UseGrid) block to prevent the messages in gridded mode. I then added a series of print statements to RunGrid which are more descriptive of the gridded mode.