SXW_Reset(void) -> SXW_Reset(char* SOILWAT_file)
(issue #300) In _run_spinup we call SXW_Reinit, which calls SXW_Reset. SXW_Reset had a reference to SXW->f_watin which is not guaranteed to exist in gridded mode. I got around this by having the function accept the file name as an input parameter. This way, it can be used in gridded or non-gridded mode.