Removed _set_sd_lyppt entirely (issue #287)
After analyzing _set_sd_lyppt I determined that the Species loop is entirely unnecessary. Once I had removed it I realized that the entire function was one line of code. It seemed unintuitive to leave a one line local function in the code so I went ahead and removed the function entirely. I then moved the functionallity into RunGrid where the call used to be.