Corrected 1-indexing issues (issue #287)
The SMort and GMort loop started at one and ended at grid_Rows and gridCols, which is 1-indexing. The loop needs to be 0-indexed to iterate through cells properly.

When creating output .csv files the formula to calculate cell number also expected 1-indexing which I had previously changed to 0-indexing, so I changed it.