Fixed historical weather file not being used.
I isolated this code from _run_spinup which makes SOILWAT use the historical weather correctly. I haven't figured out why we need to reset SXW before using it, but this seems to be an important step.

I also tried placing this code segment above _run_spinup, since I thought it couldn;t hurt and it would make the code more readable, but this didn't prevent the bug. It seems that this snippet of code must be run once; either inside _run_spinup or in RunGrid.