Created methods for saving and loading the spinup conditions.
(issue #300) This was a major endevour, and I am not yet finished.
* In ST_indivs.c I added a function for copying the contents of one individual to another individual.
* In ST_species.c I added a function for copying the contents of one SpeciesType to another SpeciesType.
* In ST_resgroups.c I added a function for copying the contents of one RGroup into another.
* In ST_grid.c I added two functions and one variable:
- spinupCells stores the state of gridCells after spinup.
- saveAsSpinupConditions() facilitates saving to spinupCells
- loadSpinupConditions() uses the copy functions above to deep copy the information from spinupCells to gridCells.

This overhaul is not finished. I need to determine what other variables need to be deep copied. For now, you can try turning on species for spinup and see that their biomass is in fact stable before the main loop starts.

NOTE: I turned UseSeedDispersal off by default. This is because seed dispersal does not work at this time.