removed use_spinup column from grid_initSpecies.csv
(issue #300) There was no reason to have this column in the file, since we can infer whether or not spinup should be used based on whether or not a species is turned on in the cell.

This commit removed deprecated error checking from _run_spinup. Since the use_spinup column no longer exists the user can never make this mistake.

This commit also introduces more error checking. If the user requests initialization with spinup but does not turn on any species for spinup they will be warned.