Overhauled soil layer swapping (issue #367)
This commit is coupled with a commit to SOILWAT2.

On the STEPWAT side I only had to add a call to set_soilLayers().  I also removed the deprecated _load_cell and _init_soil_layers. #311.

In grid_soils.in I had to modify the parameters. Transp coefficient values didn't sum to one which was crashing the program. It is unclear to me if this file had every been properly implemented, but it now contains the same inputs as soils.in from SOILWAT2.