(issue #378) Created Doxygen template
This is a basic Doxygen template. It was generated using Doxygen's built-in template generator, then modified to fit our program.  The most important difference between our configuration file and SOILWAT2's is documenting static functions. SOILWAT2 does not and ours does.

The output folder is called "Documentation" and I added it to .gitignore.