Addressing comments on PR #375
In the makefile I removed the commented out sxw_tester.o because the file is deprecated. However, I left the rule in for now in case we need the file in the future.

In ST_environs.c I added a break statement to the default case of a switch. I also put parenthesis around Plot->disturb -1 to make it clearer that this is one clause.

I removed the ternary because the first clause sufficed for setting Plot->pat_removed.