Added a main page for doxygen and a description for ST_structs.h
(issue #377)
MAinPageSetup.txt contains the information to display of the main page of the STEPWAT2 manual. I added this file to the Documents folder then added the Documents folder to the doxyfile. While I was in the doxyfile I added sw_src as well so we get all of the SOILWAT2 documentation. 

I wanted to test file descriptions so I created one for ST_structs.h