Finished documenting ST_output.c
(issue #377)
In the doxyfile I bumped the version number to reflect the latest release.

In ST_output.c I documented the file and the two functions.

I also worked on some documentation for sxw.c, but I need to wait for the changes from the gridded_code branch to finish.