Added HBAR and HMAX to inputs, created function getSpeciesHeight
(issue #422)
HBAR and HMAX are read in and stored in the SpeciesType struct. I also created the associated function, getSpeciesHeight, but for now it only returns HBAR until the actual implementation is made clear.

The values I chose for HMax need to be updated. For now they are all 40.0 centimeters.