Reconstructed function disperseSeeds
My new implementation is inline with issue #309.

I also documented ST_seedDispersal.c and .h for Doxygen.

While writing the new implementation the nested for loops were making the lines obnoxiously large (120+ characters) so I ran clang format on ST_seedDispersal.c This will make it look a little messy, but it is much easier to read in an IDE.