Created cell-specific flag for SOILWAT2 output
(issue #465)
The flag, called generateSWOutput, is declared in the CellType struct. As of now it is either TRUE for all cells or FALSE for all cells based on the -o input flag.

Spinup turns off SOILWAT2 output for obvious reasons.

The -o flag will still cause the program to crash as of this commit. The next step is to get SOILWAT2 output working.