SOILWAT2 output now generated in gridded mode
(issue #467)
My solution is, in a sense, a hack.

I guessed that the reason the -o flag was causeing a seg fault was because not enough memory was being allocated on the SOILWAT side for multiple cells. I took as evidence the fact that the program would crash 1/4 of the way through for 4 cells, 1/2 of the way through for 2 cells, and complete for 1 cell.

To combat the problem I figured out where SOILWAT2 reads the number of years that stepwat will run. I then tracked backwards into the sxw module and figured out that SXW_Reinit is the function from STEPWAT2 that tells SOIWLAT2 how many years will be run. In ST_grid.c there are two places where SXW_Reinit could be called.

I wrapped both of these functions with code that basically "tricks" SOILWAT2 into thinking that, instead of running for SuperGlobals->runModelYears, the program is running for SuperGlobals->runModelYears * grid_Cols * grid_Rows. This causes SOILWAT2 to allocate the correct amount of memory.