Created a sw_output folder in gridded mode output
Similar to nongridded mode, we need a directory present called sw_output so SOILWAT2 can put it's output there.

All I needed to add was the directory, but git will not let you commit an empty directory. I therefore put a .gitignore file in it. Technically we could delete the .gitignore file after commiting it (which appears to be the approach that non-gridded mode used) but I'm going to leave it in.

(issue #467)