Fix to reproducibly generate random number sequences
The problem was that gridded mode used the last values in a year of one cell as first values of the next cell. This went mostly unnoticed but when the last day contains precipitation, then this affects the weather generator's behavior for the first day of the next cell.

This is fixed by
(i) SOILWAT2's `SW_WTH_init_run()` did not zero out yesterday's weather values;
this was an issue only for STEPWAT2's gridded mode which does not deconstruct/construct each SOILWAT2 run (fixed with commit 8a0a7546a01b9224d85481e848c81956374d6aeb).
(ii) STEPWAT2's `load_cell()` did not call `SW_CTL_init_run()` (and `SW_WTH_init_run()`) or equivalently to prevent the carry-over of values from one cell to the next (fixed with this commit).

This now works as expected: i.e.,
* if seed != 0 (output is reproduced among runs)
** weather is exactly identical among runs and cells
** weather is different among years, iterations and seeds
* if seed == 0 (output cannot be reproduced among runs)
** weather is different among cells, years, iterations and runs

However and ideally, a grid cell should continue with the state that it ended at during the previous year (and not be zeroed out).